requirements .txt
pytest
selenium
webdriver_manager
allure-pytest
python-dotenv
mimesis